Fix check for error with empty strings
When calling this from the UI or within an action/workflow it seems that
even if you don't supply values for file or file_path, the end result is
these are getting passed as empty strings not non-existant parameters
resulting in this logic failing. Testing for empty strings fixes this
issue.